+++ pex/http.py
@@ -13,6 +13,7 @@ from .common import safe_mkdtemp, safe_open
 from .compatibility import PY3, AbstractClass
 from .tracer import TRACER
 from .variables import ENV
+from .version import __version__ as PEX_VERSION
 
 try:
   import requests
@@ -185,6 +186,7 @@ class StreamFilelike(object):
 
 class RequestsContext(Context):
   """A requests-based Context."""
+  USER_AGENT = 'pex/%s' % PEX_VERSION
 
   @staticmethod
   def _create_session(max_retries):
@@ -192,7 +194,6 @@ class RequestsContext(Context):
     retrying_adapter = requests.adapters.HTTPAdapter(max_retries=max_retries)
     session.mount('http://', retrying_adapter)
     session.mount('https://', retrying_adapter)
-
     return session
 
   def __init__(self, session=None, verify=True, env=ENV):
@@ -214,7 +215,9 @@ class RequestsContext(Context):
       return open(link.path, 'rb')  # noqa: T802
     for attempt in range(self._max_retries + 1):
       try:
-        return StreamFilelike(self._session.get(link.url, verify=self._verify, stream=True), link)
+        return StreamFilelike(self._session.get(
+            link.url, verify=self._verify, stream=True, headers={'User-Agent': self.USER_AGENT}),
+            link)
       except requests.exceptions.ReadTimeout:
         # Connect timeouts are handled by the HTTPAdapter, unfortunately read timeouts are not
         # so we'll retry them ourselves.